Add list_hud file
